requirements.txt
# LangChain
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community>=0.2.0
langchain-core>=0.2.0
langsmith>=0.1.0

# LangGraph
langgraph>=0.1.0

# MCP
mcp>=1.0.0
fastmcp>=2.0.0

# Neo4j
neo4j>=5.0.0

# FastAPI
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# SQLite / SQLAlchemy
sqlalchemy>=2.0.0

# Data
pydantic>=2.0.0
faker>=24.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utils
tenacity>=8.2.0
structlog>=24.0.0